+++ assets/js/cover.js
@@ -1,10 +1,33 @@
 var KEYCODE_ESCAPE = 27;
 
+var VIDEO_ID = 'video-overlay-video';
+
 var header_headroom;
+var player;
+
+var playButton = document.getElementById('video-overlay-play-button');
+var $featured_video = document.querySelector('#video-overlay iframe');
+
+if ($featured_video === null) {
+  $featured_video = document.querySelector('#video-overlay video');
+}
 
 function closeOverlay() {
 	jQuery('html').removeClass('noscroll');
 	jQuery('.overlay.show').removeClass('show');
+
+	// Stop playing embedded featured video, if it exists
+	if (typeof player !== 'undefined') {
+		// Youtube
+		if (player.pauseVideo) {
+			player.pauseVideo();
+		}
+
+		// Vimeo and <video> tag
+		if (player.pause) {
+			player.pause();
+		}
+	}
 }
 
 jQuery(document).ready(function() {
@@ -97,7 +120,7 @@ jQuery(document).ready(function() {
 	 * Find children by traversing up.
 	 * Not all menus or heirarchy widgets' parents has a class indicating so.
 	 */
-	jQuery('.widget .sub-menu, .widget .children, .menu .children').addClass('hide').closest('li').addClass('menu-has-child').append('<div class="menu-toggle"><i class="fa fa-angle-down"></i></div>');
+	jQuery('.widget .sub-menu, .widget .children, .menu .children, .menu .sub-menu').addClass('hide').closest('li').addClass('menu-has-child').append('<div class="menu-toggle"><i class="fa fa-angle-down"></i></div>');
 
   // click event on submenu toggles
   jQuery('body').on('click', '.menu-toggle', function(e) {
@@ -175,4 +198,76 @@ jQuery(document).ready(function() {
  */
 function isTouchDevice() {
 	return !!('ontouchstart' in window || navigator.msMaxTouchPoints);
+}
+
+if ($featured_video && $featured_video !== null) {
+  $featured_video.id = VIDEO_ID;
+
+  var sep = '?';
+  if ($featured_video.src.indexOf('?') !== -1) {
+    sep = '&';
+  }
+
+  if ($featured_video.src.indexOf('youtube.com') !== -1) {
+    // Append the embed url with the api param
+    $featured_video.src += sep + 'enablejsapi=1';
+
+    /**
+     * Inject YouTube API script
+     */
+    var tag = document.createElement('script');
+    tag.src = 'https://www.youtube.com/player_api';
+    var firstScriptTag = document.getElementsByTagName('script')[0];
+    firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
+  } else if ($featured_video.src.indexOf('vimeo.com') !== -1 && typeof Vimeo !== 'undefined') {
+    // Append the embed url with the api param
+    $featured_video.src += sep + 'api=1';
+
+    /* jshint ignore:start */
+    player = new Vimeo.Player($featured_video);
+    /* jshint ignore:end */
+
+    // Click listener for play button
+    playButton.addEventListener('click', function() {
+      /* jshint ignore:start */
+      player.play();
+      /* jshint ignore:end */
+    });
+  } else if ($featured_video.play) {
+    // Click listener for play button
+    playButton.addEventListener('click', function() {
+      $featured_video.play();
+    });
+  }
+}
+
+/**
+ * Function called once the Youtube API loads.
+ * This is dependent on checking the "Enable JavaScript API"
+ * checkbox in the plugin options (/wp-admin/options-media.php).
+ */
+function onYouTubePlayerAPIReady() {
+  // Define player as Youtube player
+  /* jshint ignore:start */
+  player = new YT.Player(VIDEO_ID, {
+    events: {
+      'onReady': onPlayerReady
+    }
+  });
+  /* jshint ignore:end */
+}
+
+/**
+ * Function to be called once the player is ready
+ */
+function onPlayerReady(event) {
+  // Get play button element
+  var playButton = document.getElementById('video-overlay-play-button');
+
+  // Click listener for play button
+  playButton.addEventListener('click', function() {
+    /* jshint ignore:start */
+    player.playVideo();
+    /* jshint ignore:end */
+  });
 }