+++ template-parts/cover-featured.php
@@ -7,9 +7,9 @@
 
 ?>
 
-<div class="featured-container">
+<div class="slider-container">
 
-	<ul class="featured-list">
+	<ul class="slider-list">
 
 	<?php
 		$img = '';
@@ -19,9 +19,12 @@
 			$img = cover_get_featured_image( $post->ID );
 	?>
 
-		<li class="featured-list-item">
+		<li class="slider-list-item">
 			<a href="<?php the_permalink(); ?>" rel="bookmark">
 				<div class="cover<?php if ( '' != $img ) { ?> featured-image<?php } ?>">
+					<?php if ( function_exists('has_post_video') && has_post_video() ) { ?>
+						<span class="svg-icon"><?php get_template_part( 'dist/svg/play', 'circle.svg' ); ?></span>
+					<?php } ?>
 					<div class="cover-background"<?php if ( '' != $img ) { ?> style="background-image: url('<?php echo $img; ?>');"<?php } ?>></div>
 					<header class="cover-header">
 						<h2 class="cover-subtitle">Featured</h2>