+++ yarpp-template-photoblog.php
@@ -1,0 +1,8 @@
+<?php
+/*
+YARPP Template: Cover Full-Width Slider
+Description: Related posts full-width slider built for the Cover theme.
+@link https://wordpress.org/themes/cover/
+Author: Paul Eiche
+*/
+require get_template_directory() . '/template-parts/yarpp-template-slider-full.php';