+++ yarpp-template-thumbnail.php
@@ -1,0 +1,8 @@
+<?php
+/*
+YARPP Template: Cover Inline List
+Description: Related posts template built for the Cover theme. Displays related posts as a comma-separated list.
+@link https://wordpress.org/themes/cover/
+Author: Paul Eiche
+*/
+require get_template_directory() . '/template-parts/yarpp-template-list.php';